experimental/camera_feed/requirements.txt
bidict==0.23.1
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.1.0
cycler==0.11.0
EasyProcess==1.1
entrypoint2==1.1
fonttools==4.42.0
h11==0.14.0
idna==3.6
imutils==0.5.4
joblib==1.3.2
kiwisolver==1.4.4
matplotlib==3.7.2
MouseInfo==0.1.3
mss==9.0.1
netifaces==0.10.6
numpy==1.25.2
opencv-contrib-python==4.6.0.66
opencv-python==4.9.0.80
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc-core==10.0
pyobjc-framework-Cocoa==10.0
pyobjc-framework-Quartz==10.0
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
python-dateutil==2.8.2
python-engineio==4.9.0
python-socketio==5.11.1
pytweening==1.0.7
pytz==2023.3
requests==2.31.0
rubicon-objc==0.4.7
scikit-learn==1.3.0
scipy==1.11.2
simple-websocket==1.0.0
six==1.16.0
socketio==0.2.1
threadpoolctl==3.2.0
tzdata==2023.3
urllib3==2.2.1
websockets==12.0
wsproto==1.2.0
